Handle window resizing
This also fixes an issue where the frame index was incorrectly used as the image index when retrieving frame buffers for rendering and moves swapchain synchronization entirely within `renderer::swapchain`. Not sure if hiding synchronization like this is a good idea, but it improves encapsulation, so I'll go with it for now.