Begin transition to a more alloc-permissive UI model
Previous code was very strict about always making use of linear-ish allocators for everything UI related. The problem with that is the poor ergonomics that I could not figure out how to fix. This solution (wherein layouts are laid out more-or-less linearly) whilst individual `RetainedElement`s aren't is a bit of a hybrid solution that I hope will work out well.

The big concern is with small `Vec`s everywhere, though the expected immediate-mode component should mean that there are relatively few arrays scattered about. It might be worth making some components such as `XSplitPanel` where only a small number of children makes sense use a fixed-capacity `ArrayVec` of some sort.

Next step is to reimplement `XSplitPanel` and then get click-and-drag resizing working for it.